requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-appconf==1.0.2
django-crispy-forms==1.7.2
django-imagekit==4.0.2
django-multiupload==0.5.2
docopt==0.6.2
idna==2.7
pilkit==2.0
Pillow==5.1.0
psycopg2==2.7.5
pytz==2018.4
requests==2.19.1
six==1.11.0
urllib3==1.23
yarg==0.1.9
